+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Fix `logger.exception` with no exception info throwing error
 ([#1006](https://github.com/census-instrumentation/opencensus-python/pull/1006))
+- Add `enable_local_storage` to turn on/off local storage + retry + flushing logic
+([#1006](https://github.com/census-instrumentation/opencensus-python/pull/1006))
 
 ## 1.0.7
 Released 2021-01-25