+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -7,7 +7,7 @@
 - Add `enable_local_storage` to turn on/off local storage + retry + flushing logic
 ([#1016](https://github.com/census-instrumentation/opencensus-python/pull/1016))
 - Enable AAD authorization via TokenCredential
-([#1016](https://github.com/census-instrumentation/opencensus-python/pull/1016))
+([#1021](https://github.com/census-instrumentation/opencensus-python/pull/1021))
 
 ## 1.0.7
 Released 2021-01-25